feat: início da simulação

+++ TP2/include/node.hpp
@@ -4,7 +4,7 @@
 
 class node {
     private:
-        int item;
+        long long int item;
         node *next;
